Fix MSVC ambiguous call error

+++ json.c
@@ -632,7 +632,7 @@ json_value * json_parse_ex (json_settings * settings, const json_char * json, ch
                         goto e_failed;
                      }
 
-                     top->u.dbl += ((double) num_fraction) / (pow (10, num_digits));
+                     top->u.dbl += ((double) num_fraction) / (pow (10, (double) num_digits));
                   }
 
                   if (b == 'e' || b == 'E')
@@ -658,7 +658,7 @@ json_value * json_parse_ex (json_settings * settings, const json_char * json, ch
                      goto e_failed;
                   }
 
-                  top->u.dbl *= pow (10, flags & flag_num_e_negative ? - num_e : num_e);
+                  top->u.dbl *= pow (10, (double) (flags & flag_num_e_negative ? - num_e : num_e));
                }
 
                if (flags & flag_num_negative)
